# schemas/dashboard-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Frontend Regression Dashboard",
    "description": "Local-first dashboard for managing UI regression runs. See REFACTOR_AND_DASHBOARD_PLAN.md milestone C for scope.",
    "version": "0.1.0"
  },
  "paths": {
    "/api/sites": {
      "get": {
        "tags": [
          "sites"
        ],
        "summary": "Get Sites",
        "description": "Read sites from `test_ui/sites.yml`.",
        "operationId": "get_sites_api_sites_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SiteOut"
                  },
                  "type": "array",
                  "title": "Response Get Sites Api Sites Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "sites"
        ],
        "summary": "Post Sites",
        "description": "Append a new site. The id is auto-generated server-side from the\nslugified name with a `-N` suffix on collision (so id-conflict 409s\nare structurally impossible - see `add_site`). Atomic write\npreserves operator-authored YAML comments.",
        "operationId": "post_sites_api_sites_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SiteCreateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteOut"
                }
              }
            }
          },
          "422": {
            "description": "Body validation failed"
          }
        }
      }
    },
    "/api/sites/{site_id}": {
      "patch": {
        "tags": [
          "sites"
        ],
        "summary": "Patch Site",
        "description": "Mutate name and/or url. id is immutable - use DELETE + POST to rename.",
        "operationId": "patch_site_api_sites__site_id__patch",
        "parameters": [
          {
            "name": "site_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Site Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SiteUpdateIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteOut"
                }
              }
            }
          },
          "404": {
            "description": "No site with this id"
          },
          "422": {
            "description": "Body validation failed"
          }
        }
      },
      "delete": {
        "tags": [
          "sites"
        ],
        "summary": "Delete Site Route",
        "description": "Remove a site from sites.yml. On-disk per-site data dirs are NOT\ntouched - historical artifacts remain readable; only future runs stop\nincluding this site.",
        "operationId": "delete_site_route_api_sites__site_id__delete",
        "parameters": [
          {
            "name": "site_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Site Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "No site with this id"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/runs": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Get Runs",
        "description": "List runs, newest first. `total` reflects the filtered set.",
        "operationId": "get_runs_api_runs_get",
        "parameters": [
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by kind",
              "title": "Kind"
            },
            "description": "Filter by kind"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by status",
              "title": "Status"
            },
            "description": "Filter by status"
          },
          {
            "name": "date_dir",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter to runs whose date_dir matches (DD-MM-YYYY). Lets the Reports page list runs for a single date without client-side truncation.",
              "title": "Date Dir"
            },
            "description": "Filter to runs whose date_dir matches (DD-MM-YYYY). Lets the Reports page list runs for a single date without client-side truncation."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunListOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "Post Runs",
        "description": "Spawn a new run as a subprocess. Returns 202 with the assigned ids.",
        "operationId": "post_runs_api_runs_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/BaselineRunRequest"
                  },
                  {
                    "$ref": "#/components/schemas/CurrentRunRequest"
                  },
                  {
                    "$ref": "#/components/schemas/ComparatorRunRequest"
                  },
                  {
                    "$ref": "#/components/schemas/ReportRunRequest"
                  }
                ],
                "discriminator": {
                  "propertyName": "kind",
                  "mapping": {
                    "baseline": "#/components/schemas/BaselineRunRequest",
                    "current": "#/components/schemas/CurrentRunRequest",
                    "comparator": "#/components/schemas/ComparatorRunRequest",
                    "report": "#/components/schemas/ReportRunRequest"
                  }
                },
                "title": "Request"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunSpawnedOut"
                }
              }
            }
          },
          "409": {
            "description": "Conflicting in-flight run for the same kind+date"
          },
          "412": {
            "description": "Workflow precondition not met (upstream not complete)"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{db_id}": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Get Run By Id",
        "operationId": "get_run_by_id_api_runs__db_id__get",
        "parameters": [
          {
            "name": "db_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Db Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunRow"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BaselineRunRequest": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "baseline",
            "title": "Kind"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind"
        ],
        "title": "BaselineRunRequest"
      },
      "ComparatorRunRequest": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "comparator",
            "title": "Kind"
          },
          "baseline_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseline Run Id"
          },
          "current_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Run Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind"
        ],
        "title": "ComparatorRunRequest"
      },
      "CurrentRunRequest": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "current",
            "title": "Kind"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind"
        ],
        "title": "CurrentRunRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReportRunRequest": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "report",
            "title": "Kind"
          },
          "date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind"
        ],
        "title": "ReportRunRequest"
      },
...
      "RunListOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RunRow"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "RunListOut",
        "description": "Paginated list response."
      },
      "RunRow": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "kind": {
            "type": "string",
            "enum": [
              "baseline",
              "current",
              "comparator",
              "report"
            ],
            "title": "Kind"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "running",
              "done",
              "failed",
              "interrupted"
            ],
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "date_dir": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Dir"
          },
          "args": {
            "additionalProperties": true,
            "type": "object",
            "title": "Args"
          },
          "command": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Command"
          },
          "exit_code": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exit Code"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "pid": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pid"
          },
          "pgid": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pgid"
          },
          "pid_start_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pid Start Time"
          },
          "source": {
            "type": "string",
            "enum": [
              "dashboard",
              "discovered",
              "cli"
            ],
            "title": "Source"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "run_id",
          "kind",
          "status",
          "created_at",
          "source"
        ],
        "title": "RunRow",
        "description": "One row from `runs`, projected for the wire.\n\n`args_json` and `command_json` are exposed as parsed objects (not raw\nstrings) - saves the frontend a JSON.parse and gives openapi-typescript\na typed shape to work with."
      },
      "RunSpawnedOut": {
        "properties": {
          "db_id": {
            "type": "integer",
            "title": "Db Id"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "running",
              "done",
              "failed",
              "interrupted"
            ],
            "title": "Status"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "db_id",
          "run_id",
          "status"
        ],
        "title": "RunSpawnedOut",
        "description": "202 response from `POST /api/runs`. `db_id` lets the frontend poll\n`/api/runs/{db_id}` for status; `run_id` lets it correlate against the\non-disk run dir as soon as the subprocess produces it."
      },
      "SiteCreateIn": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "url": {
            "type": "string",
            "maxLength": 2048,
            "minLength": 1,
            "title": "Url"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "url"
        ],
        "title": "SiteCreateIn",
        "description": "`POST /api/sites` body. `id` is auto-generated server-side from the\nslugified name + numeric dedup suffix; the client never sets it.\n\n`max_length` caps mirror the underlying `Site` model - defending at\nthe wire boundary too means a 100KB URL gets rejected with a clear\n422 instead of bloating sites.yml at the persistence layer."
      },
      "SiteOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "url": {
            "type": "string",
            "title": "Url"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name",
          "url"
        ],
        "title": "SiteOut",
        "description": "A site as returned by `GET /api/sites`."
      },
      "SiteUpdateIn": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2048,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "SiteUpdateIn",
        "description": "`PATCH /api/sites/{id}` body. Both fields optional - operator can\nrename without changing the URL or vice versa. `id` is immutable\n(changing it would invalidate existing per-site data dirs); the URL\npath component is authoritative."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}